feat: add layouts props for seo
