+++ src/pages/contact.tsx
@@ -4,7 +4,7 @@ import { NextPage } from 'next';
 import { contacts } from '~/data';
 import emailjs from 'emailjs-com';
 import { motion } from 'framer-motion';
-import Layout from '~/layouts/default';
+import Layout from '~/layouts/defaultLayout';
 import ContactForm from '~/components/ContactForm';
 import { useToasts } from 'react-toast-notifications';
 import { stagger, fadeInUp } from '~/animation';
@@ -41,64 +41,57 @@ const Contact: NextPage<ContactPageProps> = () => {
   };
 
   return (
-    <>
-      <Head>
-        <title>Contact | Joshua Galit</title>
-        <meta name="description" content="Contact through email or phone" />
-      </Head>
-      <Layout>
-        <div className="w-full max-w-5xl m-auto px-4 py-4 md:py-0">
-          <div className="w-full h-full opacity-30 absolute">
-            <Image
-              src="/svgs/buble.svg"
-              layout="fill"
-              alt="Bubble Background"
-            />
-          </div>
-          <div className="h-screen min-h-screen flex flex-col md:flex-row items-start md:items-center justify-between space-y-8">
-            <motion.div
-              variants={stagger}
-              className="flex flex-col space-y-10 md:space-y-20"
-            >
-              <motion.div
-                variants={fadeInUp}
-                className="flex flex-col space-y-5 md:space-y-8 w-full max-w-md z-50"
-              >
-                <h1 className="font-black text-3xl md:text-5xl text-blue-twitter">
-                  Contact me
-                </h1>
-                <p className="text-base text-gray-900 dark:text-white">
-                  Send me a message and I will get back to you within 24 hours.
-                </p>
-              </motion.div>
-              <motion.div
-                variants={stagger}
-                className="flex flex-col space-y-3 md:space-y-5 z-50"
-              >
-                {contacts.map(({ Icon, text }, i) => (
-                  <motion.div
-                    key={i}
-                    variants={fadeInUp}
-                    className="flex items-center space-x-6 text-gray-900 dark:text-white"
-                  >
-                    <Icon className="w-6 h-6 text-blue-twitter" />
-                    <p className="text-base">{text}</p>
-                  </motion.div>
-                ))}
-              </motion.div>
-            </motion.div>
+    <Layout
+      headTitle="Contact | Joshua Galit"
+      metaContent="Contact through email or phone"
+    >
+      <div className="w-full max-w-5xl m-auto px-4 py-4 md:py-0">
+        <div className="w-full h-full opacity-30 absolute">
+          <Image src="/svgs/buble.svg" layout="fill" alt="Bubble Background" />
+        </div>
+        <div className="h-screen min-h-screen flex flex-col md:flex-row items-start md:items-center justify-between space-y-8">
+          <motion.div
+            variants={stagger}
+            className="flex flex-col space-y-10 md:space-y-20"
+          >
             <motion.div
               variants={fadeInUp}
-              className="rounded-lg w-full max-w-full md:max-w-lg px-8 py-10 bg-white shadow-md dark:bg-gray-900 z-50"
+              className="flex flex-col space-y-5 md:space-y-8 w-full max-w-md z-50"
             >
-              <div className="relative w-full">
-                <ContactForm onSubmit={handleContact} />
-              </div>
+              <h1 className="font-black text-3xl md:text-5xl text-blue-twitter">
+                Contact me
+              </h1>
+              <p className="text-base text-gray-900 dark:text-white">
+                Send me a message and I will get back to you within 24 hours.
+              </p>
             </motion.div>
-          </div>
+            <motion.div
+              variants={stagger}
+              className="flex flex-col space-y-3 md:space-y-5 z-50"
+            >
+              {contacts.map(({ Icon, text }, i) => (
+                <motion.div
+                  key={i}
+                  variants={fadeInUp}
+                  className="flex items-center space-x-6 text-gray-900 dark:text-white"
+                >
+                  <Icon className="w-6 h-6 text-blue-twitter" />
+                  <p className="text-base">{text}</p>
+                </motion.div>
+              ))}
+            </motion.div>
+          </motion.div>
+          <motion.div
+            variants={fadeInUp}
+            className="rounded-lg w-full max-w-full md:max-w-lg px-8 py-10 bg-white shadow-md dark:bg-gray-900 z-50"
+          >
+            <div className="relative w-full">
+              <ContactForm onSubmit={handleContact} />
+            </div>
+          </motion.div>
         </div>
-      </Layout>
-    </>
+      </div>
+    </Layout>
   );
 };
